Header files moved to include/geos. Header files are being installed now.
