+++ include/geos/EnumString.h
