+++ include/geos/Profiling/GEOSProfiler.h
@@ -12,7 +12,7 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#include "ProfileModule/ProfileModule.h"
+#include "geos/ProfileModule/ProfileModule.h"
 #include "llvm/IR/Module.h"
 
 class GEOSProfiler {
