+++ src/CostEstimator/CostAnalysis/Branch/BranchAnalysis.cpp
@@ -12,9 +12,9 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#include "CostEstimator/CostAnalysis.h"
+#include "geos/CostEstimator/CostAnalysis.h"
 
-#include "CostEstimator/InstructionCostEstimator.h"
+#include "geos/CostEstimator/InstructionCostEstimator.h"
 
 using namespace llvm;
