+++ src/CostEstimator/CostAnalysis/Call/CallCostAnalysis.cpp
@@ -15,8 +15,8 @@
 
 #include "llvm/IR/Module.h"
 
-#include "CostEstimator/CostAnalysis.h"
-#include "CostEstimator/InstructionCostEstimator.h"
+#include "geos/CostEstimator/CostAnalysis.h"
+#include "geos/CostEstimator/InstructionCostEstimator.h"
 
 using namespace llvm;
