+++ src/Profiling/ExecutionFrequency/Static/StaticProfiling.cpp
@@ -13,7 +13,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "Profiling/StaticProfiling.h"
+#include "geos/Profiling/StaticProfiling.h"
 
 #include "llvm/Analysis/BlockFrequencyInfo.h"
 #include "llvm/Analysis/CallGraph.h"
