+++ tools/CodeMetric.cpp
@@ -12,7 +12,7 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#include "GEOS.h"
+#include "geos/GEOS.h"
 
 #include "llvm/Bitcode/ReaderWriter.h"
 #include "llvm/IRReader/IRReader.h"   
@@ -21,8 +21,8 @@
 #include "llvm/Support/SourceMgr.h"
 #include "llvm/IR/LLVMContext.h"
 
-#include "ProfileModule/ModuleMetric.h"
-#include "GEOSCommandLineParser.h"
+#include "geos/ProfileModule/ModuleMetric.h"
+#include "geos/GEOSCommandLineParser.h"
 
 #include <vector>
 #include <limits>
@@ -31,9 +31,9 @@
 #include <string>
 #include <fstream>
 
-#include "OfflineLearning/CodeMetricBase.h"
-#include "OfflineLearning/AccuracyBase.h"
-#include "OfflineLearning/CorrectionBase.h"
+#include "geos/OfflineLearning/CodeMetricBase.h"
+#include "geos/OfflineLearning/AccuracyBase.h"
+#include "geos/OfflineLearning/CorrectionBase.h"
 
 using namespace llvm;
 using std::string;
